+++ src/app/src/redux/app.actions.js
@@ -5,3 +5,4 @@ export const setGeoSelectionMode = createAction('Set geo selection mode');
 export const setGeoSelection = createAction('Set geo selection');
 export const setQuestionKeys = createAction('Set question keys');
 export const setVisualization = createAction('Set visualization');
+export const setYears = createAction('Set years');