+++ src/app/src/components/QuestionSelector.jsx
@@ -22,7 +22,12 @@ import { IconContext } from 'react-icons';
 
 import { setQuestionKeys } from '../redux/app.actions';
 import { CONFIG, ROUTES } from '../utils/constants';
-import { formatQuery, DataIndexer } from '../utils';
+import {
+    formatQuery,
+    DataIndexer,
+    calculateAvailableGeo,
+    formatCurrentGeo,
+} from '../utils';
 import Breadcrumbs from './Breadcrumbs';
 import SearchInput from './SearchInput';
 
@@ -44,6 +49,13 @@ const QuestionSelector = () => {
     // Select the appropriate config file based on the current geoMode
     const config = CONFIG[geoMode];
 
+    // Select the appropriate config file based on the current geoMode
+    const survey = config?.survey;
+    const years = useMemo(
+        () => (geoMode && data[geoMode] ? Object.keys(data[geoMode]) : []),
+        [geoMode, data]
+    );
+
     // Determine the indexes of any categories that currently have questions
     // selected. These will be expanded by default.
     const categoryLetterCodes = Object.values(config.categories);
@@ -62,6 +74,13 @@ const QuestionSelector = () => {
         if (!currentGeo.length || !currentYears.length) return null;
         return new DataIndexer(currentYears, geoMode, currentGeo, data);
     }, [currentYears, geoMode, currentGeo, data]);
+
+    // Select the available years based on available questions for selected geographies
+    const availableYearsGeography = useMemo(
+        () =>
+            calculateAvailableGeo({ years, geoMode, currentGeo, data, survey }),
+        [years, geoMode, currentGeo, data, survey]
+    );
 
     // Known categories
     const questionsByCategory = { A: [], B: [], C: [], D: [] };
@@ -372,7 +391,11 @@ const QuestionSelector = () => {
                     <Box as='span' opacity='0.5' mx={1}>
                         •
                     </Box>
-                    {currentGeo.join(', ')}
+                    {formatCurrentGeo({
+                        currentGeo,
+                        currentYears,
+                        availableYearsGeography,
+                    })}
                 </Text>
             </Flex>
             <Flex