+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Update get-pip url [#109](https://github.com/azavea/fb-gender-survey-dashboard/pull/109)
+
 ### Removed
 
 # 1.4.0