+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Enable cross-origin requests [#111](https://github.com/azavea/fb-gender-survey-dashboard/pull/111)
 - Use local data for development [#110](https://github.com/azavea/fb-gender-survey-dashboard/pull/110)
+- Add year selector [#116](https://github.com/azavea/fb-gender-survey-dashboard/pull/116)
 
 ### Fixed