+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Process 2021 config data [#120](https://github.com/azavea/fb-gender-survey-dashboard/pull/120)
+
 ### Changed
 
 - Enable cross-origin requests [#111](https://github.com/azavea/fb-gender-survey-dashboard/pull/111)